feat(exercise): Implementa CRUD completo para Exercícios

+++ src/main/java/io/github/matheusplaza/fitlogtds/repository/ExerciseRepository.java
@@ -3,5 +3,5 @@ package io.github.matheusplaza.fitlogtds.repository;
 import io.github.matheusplaza.fitlogtds.model.Exercise;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface ExerciseRepository extends JpaRepository<Exercise, Integer> {
+public interface ExerciseRepository extends JpaRepository<Exercise, Long> {
 }
